Changed values from int array to float array to support decimals

+++ sort-array/src/Main.java
@@ -2,12 +2,12 @@ import java.util.Arrays;
 
 public class Main {
     public static void main(String[] args) {
-        int[] values = {20, 20, 20, 32, 36, 37, 39, 43, 46, 47, 48, 50, 51, 51, 62, 64, 65};
+        float[] values = {20, 20, 20, 32, 36, 37, 39, 43, 46, 47, 48, 50, 51, 51, 62, 64, 65};
 
         Arrays.sort(values);
         int count = 0;
 
-        for (int num : values) {
+        for (float num : values) {
             count++;
             System.out.print(num);
